Separate causal DirectoryPreparer for easier merges
