+++ corebehrt/main_causal/helper/calibrate_mlp.py
@@ -9,8 +9,10 @@ from torch.utils.data import DataLoader
 
 from corebehrt.constants.causal.data import CF_PROBAS, PROBAS, TARGETS
 from corebehrt.constants.data import PID_COL
-from corebehrt.main_causal.helper.calibrate import (calibrate_probas,
-                                                    train_isotonic_regression)
+from corebehrt.main_causal.helper.calibrate import (
+    calibrate_probas,
+    train_isotonic_regression,
+)
 
 
 def collect_predictions(
