+++ corebehrt/main_causal/simulate.py
@@ -14,7 +14,7 @@ from corebehrt.functional.setup.args import get_args
 from corebehrt.functional.utils.time import get_hours_since_epoch
 from corebehrt.main_causal.helper.simulate import simulate
 from corebehrt.modules.setup.config import load_config
-from corebehrt.modules.setup.directory import DirectoryPreparer
+from corebehrt.modules.setup.directory_causal import CausalDirectoryPreparer
 
 CONFIG_PATH = "./corebehrt/configs/causal/simulate.yaml"
 
@@ -23,7 +23,7 @@ def main_simulate(config_path):
     cfg = load_config(config_path)
 
     # Setup directories
-    DirectoryPreparer(cfg).setup_simulate()
+    CausalDirectoryPreparer(cfg).setup_simulate()
 
     # Logger
     logger = logging.getLogger("simulate")
