+++ corebehrt/modules/setup/directory.py
@@ -4,24 +4,16 @@ import uuid
 from os.path import basename, join, splitext
 from shutil import copyfile, rmtree
 
-from corebehrt.constants.causal.paths import (
-    ENCODE_CFG,
-    SIMULATE_CFG,
-    CALIBRATE_CFG,
-    TRAIN_MLP_CFG,
-    ESTIMATE_CFG,
-)
 from corebehrt.constants.paths import (
+    CHECKPOINTS_DIR,
+    COHORT_CFG,
+    DATA_CFG,
     FINETUNE_CFG,
     OUTCOMES_CFG,
     PRETRAIN_CFG,
     PREPARE_PRETRAIN_CFG,
     PREPARE_FINETUNE_CFG,
-    CHECKPOINTS_DIR,
-    COHORT_CFG,
-    DATA_CFG,
 )
-
 from corebehrt.functional.setup.checks import check_categories
 from corebehrt.modules.setup.config import Config, load_config
 
@@ -376,103 +368,6 @@ class DirectoryPreparer:
         if "tokenized" not in self.cfg.paths:
             logger.info("Tokenized dir not in config. Adding from pretrain config.")
             self.cfg.paths.tokenized = data_cfg.paths.tokenized
-
-    def setup_encode(self) -> None:
-        """
-        Validates path config and sets up directories for encode.
-        """
-        # Setup logging
-        self.setup_logging("encode")
-
-        # Validate and create directories
-        self.check_directory("finetune_model")
-        self.create_directory("encoded_data")
-
-        # Write config in output directory.
-        self.write_config("encoded_data", source="finetune_model", name=PRETRAIN_CFG)
-        self.write_config("encoded_data", source="finetune_model", name=FINETUNE_CFG)
-        self.write_config("encoded_data", name=ENCODE_CFG)
-
-    def setup_simulate(self) -> None:
-        """
-        Validates path config and sets up directories for simulate.
-        """
-        # Setup logging
-        self.setup_logging("simulate")
-
-        # Validate and create directories
-        self.check_directory("calibrated_predictions")
-        self.check_directory("encoded_data")
-        self.create_directory("simulated_outcome")
-
-        # Write config in output directory.
-        self.write_config(
-            "simulated_outcome", source="calibrated_predictions", name=CALIBRATE_CFG
-        )
-        self.write_config(
-            "simulated_outcome", source="calibrated_predictions", name=FINETUNE_CFG
-        )
-        self.write_config("simulated_outcome", source="encoded_data", name=ENCODE_CFG)
-        self.write_config("simulated_outcome", name=SIMULATE_CFG)
-
-    def setup_calibrate(self) -> None:
-        """
-        Validates path config and sets up directories for calibrate.
-        """
-        # Setup logging
-        self.setup_logging("calibrate")
-
-        # Validate and create directories
-        self.check_directory("finetune_model")
-        self.create_directory("calibrated_predictions")
-
-        # Write config in output directory.
-        self.write_config(
-            "calibrated_predictions", source="finetune_model", name=FINETUNE_CFG
-        )
-        self.write_config("calibrated_predictions", name=CALIBRATE_CFG)
-
-    def setup_train_mlp(self) -> None:
-        """
-        Validates path config and sets up directories for train_mlp.
-        """
-        # Setup logging
-        self.setup_logging("train_mlp")
-
-        # Validate and create directories
-        self.check_directory("encoded_data")
-        self.check_directory("calibrated_predictions")
-        self.check_directory("cohort")
-        self.create_run_directory("trained_mlp", base="runs")
-
-        # Write config in output directory.
-        self.write_config("trained_mlp", source="encoded_data", name=ENCODE_CFG)
-        self.write_config(
-            "trained_mlp", source="calibrated_predictions", name=CALIBRATE_CFG
-        )
-        self.write_config("trained_mlp", source="cohort", name=COHORT_CFG)
-        self.write_config("trained_mlp", name=TRAIN_MLP_CFG)
-
-    def setup_estimate(self) -> None:
-        """
-        Validates path config and sets up directories for estimate.
-        """
-        # Setup logging
-        self.setup_logging("estimate")
-
-        # Validate and create directories
-        self.check_directory("exposure_predictions")
-        self.check_directory("outcome_predictions")
-
-        # Optional counterfactual outcomes check
-        if self.cfg.paths.get("counterfactual_outcomes", False):
-            self.check_directory("counterfactual_outcomes")
-
-        # Create estimate directory
-        self.create_run_directory("estimate", base="runs")
-
-        # Write config in output directory
-        self.write_config("estimate", name=ESTIMATE_CFG)
 
     #
     # Directory naming generators
